Initial commit of the node features
We need to implement common behaviour through the concept
of "features". The main goal is to be able to make efl-egueb
create windows for the render/ui features. For animation feature
we can also make ti work with ecore and so on. So later creating
eon based windows will be simpler.