Look for the inner use elements for element_at
Basically the use element has a shadow tree of svg elements
The spec says that those elements must be treated as a special
object SVGElementInstance, but for now we just pass directly
the cloned element.
In case a use references some <a> elements we need to be sure
to also iterate over the inner group of the use.