Support recursive processes
Basically when an element was enqueued during the processing
of a document, it was enqueued to be processed the next time
the document is processed. We now support to process the list
until no elements are left on the queue. This has a problem
of infinite loop so we need to break the loop sometime, haven't
decided when tho